@@ -2,13 +2,13 @@
 
 buildscript {
     ext.android_support_version = '28.0.0'
-    ext.kotlin_version = '1.4.21'
-    ext.anko_version = '0.10.8'
+    ext.kotlin_version = '1.5.30'
     ext.retrofit_version = '2.9.0'
-    ext.okhttp_version = '4.9.0'
+    ext.okhttp_version = '4.9.1'
+    ext.gson_version = '2.8.8'
     repositories {
-        maven { url 'https://maven.aliyun.com/repository/google' }
-        maven { url 'https://maven.aliyun.com/repository/jcenter' }
+        google()
+        mavenCentral()
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:7.0.2'
@@ -22,8 +22,8 @@ buildscript {
 
 allprojects {
     repositories {
-        maven { url 'https://maven.aliyun.com/repository/google' }
-        maven { url 'https://maven.aliyun.com/repository/jcenter' }
+        google()
+        mavenCentral()
         maven { url 'https://jitpack.io' }
     }
 }
