@@ -4,15 +4,15 @@ apply plugin: 'kotlin-android-extensions'
 group = 'com.github.zeropercenthappy'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 30
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 28
+        targetSdkVersion 30
         versionCode 120
         versionName "1.2.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
 
         compileOptions {
             sourceCompatibility = '1.8'
@@ -36,12 +36,13 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.9"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.0"
 
-    implementation 'com.github.zeropercenthappy:ZPHAndroidUtils:1.3.6'
+    implementation 'com.github.zeropercenthappy:ZPHAndroidUtils:1.4.0'
     api "com.squareup.okhttp3:okhttp:$okhttp_version"
     api "com.squareup.retrofit2:retrofit:$retrofit_version"
     api "com.squareup.retrofit2:converter-gson:$retrofit_version"
+    api "com.google.code.gson:gson:$gson_version"
     api 'org.conscrypt:conscrypt-android:2.2.1'
 }
 
