@@ -2,15 +2,15 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 android {
-    compileSdkVersion 28
+    compileSdkVersion 30
     defaultConfig {
         applicationId "com.zeropercenthappy.retrofitutilsample"
         minSdkVersion 15
-        targetSdkVersion 27
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
 
         compileOptions {
             sourceCompatibility = '1.8'
@@ -31,17 +31,14 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation "com.android.support:appcompat-v7:$android_support_version"
-    implementation "com.android.support.constraint:constraint-layout:2.0.4"
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.9"
-    implementation("org.jetbrains.anko:anko:$anko_version") {
-        exclude group: "com.android.support"
-    }
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.0"
 
     implementation project(':retrofitLib')
-    implementation 'com.github.zeropercenthappy:OkHttpLogInterceptor:1.6.5'
-    implementation 'com.github.zeropercenthappy:ZPHAndroidUtils:1.3.6'
+    implementation 'com.github.zeropercenthappy:OkHttpLogInterceptor:1.6.7'
+    implementation 'com.github.zeropercenthappy:ZPHAndroidUtils:1.4.0'
     implementation('com.yanzhenjie:album:2.1.3')
 }
